refactor: rewrite to github api

+++ R/upload.R
@@ -11,8 +11,6 @@
 #' @param data_output_path Path where the reshaped databases will be written
 #' @param action action to be performed, can be 'reshape', 'populate' or 'all'
 #'
-#' @importFrom utils packageVersion
-#'
 #' @export
 lc.upload <-
   local(function(dict_version = '2_0',
@@ -25,10 +23,19 @@ lc.upload <-
                  data_output_path = getwd(),
                  action = "all",
                  upload_to_opal = TRUE) {
-    if (!exists('hostname', envir = lifecycle.globals))
-      stop('You need to login first, please run lc.login')
-    if (!exists('username', envir = lifecycle.globals))
-      stop('You need to login first, please run lc.login')
+    
+    message('######################################################')
+    message('  Start upload data into Opal')
+    message('------------------------------------------------------')
+    
+    populateDictionaryVersions(dict_kind)
+    
+    if(upload_to_opal == TRUE) {
+      if (!exists('hostname', envir = lifecycle.globals))
+        stop('You need to login first, please run lc.login')
+      if (!exists('username', envir = lifecycle.globals))
+        stop('You need to login first, please run lc.login')
+    }
     
     if (missing(cohort_id))
       cohort_id <-
@@ -135,8 +142,12 @@ lc.upload <-
     finally = {
       message(" * Reinstate default working directory")
       setwd(workdir)
-      message(" * Cleanup temporary directory")
-      unlink(tempDirectoryName[1], recursive = T)
+      if (upload_to_opal == TRUE) {
+        message(" * Cleanup temporary directory")
+        unlink(tempDirectoryName[1], recursive = T)
+      } else {
+        message(" * Be advised: you need to cleanup the temporary directories yourself now.")
+      }
     })
